+++ pkg/csi/utils.go
@@ -29,7 +29,7 @@ import (
 	pxapi "github.com/Telmate/proxmox-api-go/proxmox"
 	"github.com/siderolabs/go-retry/retry"
 
-	volume "github.com/sergelogvinov/proxmox-csi-plugin/pkg/volume"
+	volume "github.com/sergelogvinov/proxmox-csi-plugin/pkg/utils/volume"
 )
 
 const (
